Added tests

+++ quantum/test/Dialect/QIR/IR/ops.mlir
@@ -1,4 +1,4 @@
-// RUN: quantum-opt %s --convert-qir-to-llvm | FileCheck %s
+// RUN: quantum-opt %s --mlir-print-ir-after-all | FileCheck %s
 
 func.func @main() {
   // Allocate qubits and results.
@@ -11,12 +11,12 @@ func.func @main() {
   %r2 = "qir.ralloc" () : () -> (!qir.result)
 
   // Create constants for rotation angles.
-  %const1 = arith.constant 0.34 : f32
-  %const2 = arith.constant 0.735 : f32
+  %const1 = arith.constant 0.34 : f64
+  %const2 = arith.constant 0.735 : f64
 
   // QIR operations.
   "qir.H" (%q0) : (!qir.qubit) -> ()
-  "qir.Rz" (%q0, %const1) : (!qir.qubit, f32) -> ()
+  "qir.Rz" (%q0, %const1) : (!qir.qubit, f64) -> ()
   "qir.H" (%q1) : (!qir.qubit) -> ()
   "qir.H" (%q2) : (!qir.qubit) -> ()
   "qir.measure" (%q0, %r0) : (!qir.qubit, !qir.result) -> ()
@@ -26,13 +26,23 @@ func.func @main() {
   return
 }
 
-// CHECK-LABEL: func.func @main()
-// CHECK: %q0 = "qir.alloc" () : () -> (!qir.qubit)
-// CHECK: %const1 = arith.constant 0.34 : f32
-// CHECK: call void @__quantum__qis__h__body(%q0)
-// CHECK: call void @__quantum__qis__rz__body(%const1, %q0)
-// CHECK: call void @__quantum__qis__h__body(%q1)
-// CHECK: call void @__quantum__qis__h__body(%q2)
-// CHECK: call void @__quantum__qis__measure__body(%q0, %r0)
-// CHECK: call void @__quantum__qis__measure__body(%q1, %r1)
-// CHECK: call void @__quantum__qis__measure__body(%q2, %r2)
+// CHECK: module {
+// CHECK:   func.func @main() {
+// CHECK:     %0 = "qir.alloc"() : () -> !qir.qubit
+// CHECK:     %1 = "qir.alloc"() : () -> !qir.qubit
+// CHECK:     %2 = "qir.alloc"() : () -> !qir.qubit
+// CHECK:     %3 = "qir.ralloc"() : () -> !qir.result
+// CHECK:     %4 = "qir.ralloc"() : () -> !qir.result
+// CHECK:     %5 = "qir.ralloc"() : () -> !qir.result
+// CHECK:     %cst = arith.constant 3.400000e-01 : f64
+// CHECK:     %cst_0 = arith.constant 7.350000e-01 : f64
+// CHECK:     "qir.H"(%0) : (!qir.qubit) -> ()
+// CHECK:     "qir.Rz"(%0, %cst) : (!qir.qubit, f64) -> ()
+// CHECK:     "qir.H"(%1) : (!qir.qubit) -> ()
+// CHECK:     "qir.H"(%2) : (!qir.qubit) -> ()
+// CHECK:     "qir.measure"(%0, %3) : (!qir.qubit, !qir.result) -> ()
+// CHECK:     "qir.measure"(%1, %4) : (!qir.qubit, !qir.result) -> ()
+// CHECK:     "qir.measure"(%2, %5) : (!qir.qubit, !qir.result) -> ()
+// CHECK:     return
+// CHECK:   }
+// CHECK: }
